pyproject.toml
[project]
name = "zighang-zighang-embedding"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "numpy>=2.0.2",
    "opensearch-py>=3.0.0",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "torch>=2.8.0",
    "torchaudio",
    "torchvision",
    "transformers>=4.55.4",
]

[tool.uv.sources]
torch = { index = "pytorch-cu129" }
torchvision = { index = "pytorch-cu129" }
torchaudio = { index = "pytorch-cu129" }

[[tool.uv.index]]
name = "pytorch-cu129"
url = "https://download.pytorch.org/whl/cu129"
explicit = true
